renders exit attributes use document attributes map object CHM-852
