+++ app/app-text/views/forms/edit/abs/edit-legal-framework-overview.json
@@ -10,31 +10,33 @@
   "jurisdiction": "Jurisdiction",
   "jurisdictionImplementationNational": "If you selected National above, please provide any additional information related to the implementation of the Protocol",
   "jurisdictionImplementationSubNational": "If you selected Subnational or Community above, please indicate to how the selected jurisdiction’s application differs from the national application of the of the Protocol.",
-  "establishedMeasure": "Has your country established legislative, administrative or policy measures on ABS?",
-  "geneticResources": "Access to genetic resources",
-  "agrSubjectToPic": "Is access to genetic resources subject to prior informed consent",
-  "agrMeasureForAccess": "Are measures in place related to access to genetic resources?",
+  "establishedMeasure": "Has your country established legislative, administrative or policy measures on ABS? {establishedMeasuresInclude}",
+  "establishedMeasuresInclude": "This includes measures on access, benefit-sharing and compliance.",
+  "geneticResourcesAccess": "Access to {geneticResources}",
+  "agrSubjectToPic": "Is access to {geneticResources} subject to prior informed consent?",
+  "agrMeasureForAccess": "Are measures in place related to access to {geneticResources}?",
   "commercial": "commercial",
   "nonCommercial": "non-commercial",
-  "argPermitRequired": "Is an ABS permit required for",
-  "argPermitUtilization": "utilization of genetic resources?",
+  "argPermitRequired": "Is an ABS permit required for {type} utilization of {geneticResources}?",
   "benefitSharing": "Measures on fair and equitable benefit-sharing",
   "anyExceptions": "Are there any exceptions?",
   "yesAllMeasures": "Yes, all measures are in place",
-  "tkSubjectToPic": "Is access to traditional knowledge associated with genetic resources subject to prior informed consent?",
+  "tkSubjectToPic": "Is access to {traditionalKnowledge} subject to prior informed consent?",
   "yesAllCases": "Yes, in all cases",
   "yesSomeCases": "Yes, in some cases",
   "noCountryDetermined": "No, my country has determined that access to its genetic resources is not subject to prior informed consent",
   "knowledgeAccess": "Access to traditional knowledge associated with genetic resources",
-  "tkMeasureForAccess": "Are ABS measures in place related to access to traditional knowledge associated with genetic resources?",
+  "tkMeasureForAccess": "Are ABS measures in place related to access to {traditionalKnowledge}?",
+  "traditionalKnowledge": "traditional knowledge associated with genetic resources",
   "yesToSomeExtent": "Yes, to some extent",
-  "tkPermitRequired": "Is an ABS permit required for {msg} utilization of traditional knowledge associated with genetic resources?",
+  "tkPermitRequired": "Is an ABS permit required for {type} utilization of {traditionalKnowledge}?",
   "article53Implemented": "15. Does your country have measures in place requiring that benefits from the utilization of genetic resources, as well as subsequent applications and commercialization, be shared with the Party providing such resources?",
   "article55Implemented": "17. Has your country taken measures in order that benefits arising from the utilization of traditional knowledge associated with genetic resources are shared with indigenous peoples and local communities holding such knowledge?",
   "pleaseExplain": "Please explain:",
   "national": "National",
   "subNational":" Subnational",
   "community":" Community",
+  "geneticResources": "genetic resources",
   "no": "No",
   "yes": "Yes"
 }
